created abstraction for detectors, minor improvements
