+++ cases/anomaly_detection/clear_architecture/utils/data_types.py
