+++ cases/anomaly_detection/data/label_reader.py
@@ -1,6 +1,6 @@
 
-import sys
 import os
+import sys
 
 sys.path.append('../utils')
 import_path = os.path.dirname(os.path.abspath(__file__))
