+++ cases/classification_experiment.py
@@ -5,4 +5,3 @@ if __name__ == '__main__':
     config_name = 'ConfigAll.yaml'
     ExperimentHelper = Industrial()
     ExperimentHelper.run_experiment(config_name)
-    _ = 1
